Cleanup: various changes
Lots of small changes were made,

The weird break macros were removed and will be recomended to use
the try operator (?) inside a function that returns io::Result<()>.
The `general_event_process!` macro was also made public.

Slight changes to builder pattern so it is clear the only required input
is forced to be provided through the `repl_builder` fn (the only way to
get access to the builder struct)

One Unix render bug was fixed, this was because of a missing
`line_at_start` flag set. Unix hidden cursor on close was also fixed