Ensure all history entries are unique
History now uses a BTreeMap to maintain the order of all history
elements and a HashMap to locate the current location of any given
string inside the BtreeMap. To avoid History becoming a self-referential
struct we use the u64 generated by DefaultHasher as keys in the HashMap.
This does result in 2 hash operations per insert.