Feat: tag input hooks
It is now possible to tag input hooks, Input hooks are tagged with `i32`
values. Currently it is setup so any type that impls `Into<i32>` can be
used. This makes it very straight forward to directly pass in enums that
all have unit fields.

Tagging input hooks allows for users to remove current, and queued input
hooks dynamically. There are new methods in the public api for these
events.