Fix: possible incorrect error formatting
In the case a library user tried to execute an async callback without
clearning the current line, if that callback returned an error the call
to `tracing::error` would not display correctly. Assuming that the users
subscriber used the same output as the REPL.

`prep_for_background_msg` now internally returns early if the call is not
needed.